[gemini] update optimizer interface
